nx531j: Parts: add qcamerasvr restart workaround
Unfortunately, z11's mm-qcamera-daemon has an issue where sometimes
it can't receive any valid preview frame at times which causes the
camera to black screen and fail to work until a reboot is
performed. This issue could be due to us using camera HAL3 when
Nubia only inteded for HAL1 to be used as HAL3 is disabled on our
stock ROM.

Without the source code to mm-qcamera-daemon and the camera blobs
we cannot currently fix this issue. A workaround we can do though
to prevent a reboot is to restart the qcamerasvr manually by
adding a setting for it.

This commit is co-authored by @zhaoguomanong & @ThEMarD as they implemented it
on leeco s2